# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medical Appointment API",
    "version": "1.0.0",
    "description": "Swagger documentation for the medical appointment backend API."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "uptime": {
            "type": "number",
            "example": 12.34
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          }
        }
      },
      "SetClaimsRequest": {
        "type": "object",
        "required": [
          "uid",
          "claims"
        ],
        "properties": {
          "uid": {
            "type": "string",
            "example": "firebase-user-id"
          },
          "claims": {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "enum": [
                  "doctor",
                  "manager",
                  "user",
                  "admin"
                ],
                "example": "admin"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Validation error"
              },
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Appointment": {
        "type": "object",
        "required": [
          "id",
          "clinic",
          "type",
          "spots",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "clinic": {
            "type": "string",
            "minLength": 2,
            "example": "City Medical Center"
          },
          "type": {
            "type": "string",
            "minLength": 2,
            "example": "General Checkup"
          },
          "spots": {
            "type": "integer",
            "minimum": 0,
            "example": 3
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "open",
              "full",
              "delayed",
              "cancelled"
            ],
            "example": "open"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-10T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-10T10:00:00.000Z"
          }
        }
      },
      "CreateAppointmentInput": {
        "type": "object",
        "required": [
          "clinic",
          "type",
          "spots",
          "status"
        ],
        "properties": {
          "clinic": {
            "type": "string",
            "minLength": 2,
            "example": "North Clinic"
          },
          "type": {
            "type": "string",
            "minLength": 2,
            "example": "Vaccination"
          },
          "spots": {
            "type": "integer",
            "minimum": 0,
            "example": 4
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "open",
              "full",
              "delayed",
              "cancelled"
            ],
            "example": "open"
          }
        }
      },
      "WrappedCreateAppointmentInput": {
        "type": "object",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/CreateAppointmentInput"
          }
        }
      },
      "UpdateAppointmentInput": {
        "type": "object",
        "description": "At least one appointment field must be provided.",
        "properties": {
          "clinic": {
            "type": "string",
            "minLength": 2,
            "example": "Updated Clinic"
          },
          "type": {
            "type": "string",
            "minLength": 2,
            "example": "Follow-up"
          },
          "spots": {
            "type": "integer",
            "minimum": 0,
            "example": 2
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "open",
              "full",
              "delayed",
              "cancelled"
            ],
            "example": "full"
          }
        }
      },
      "WrappedUpdateAppointmentInput": {
        "type": "object",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/UpdateAppointmentInput"
          }
        }
      },
      "AppointmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Appointment created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Appointment"
          }
        }
      },
      "AppointmentListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Appointment"
        }
      }
    }
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "summary": "Health check",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "API health information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/setClaims": {
      "post": {
        "summary": "Set Firebase custom claims for a user",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetClaimsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Claims updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Ok"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apts": {
      "get": {
        "summary": "Get all appointments",
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "tags": [
          "Appointments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateAppointmentInput"
                  },
                  {
                    "$ref": "#/components/schemas/WrappedCreateAppointmentInput"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apts/{id}": {
      "get": {
        "summary": "Get an appointment by id",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      },
      "put": {
        "summary": "Update an appointment by id",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateAppointmentInput"
                  },
                  {
                    "$ref": "#/components/schemas/WrappedUpdateAppointmentInput"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "delete": {
        "summary": "Delete an appointment by id",
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "418": {
            "description": "Appointment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}